Testsuite test: Moved TH stuff to own modules
