+++ path.cabal
@@ -56,6 +56,8 @@ test-suite test
                    , Windows
                    , Common.Posix
                    , Common.Windows
+                   , TH.Posix
+                   , TH.Windows
   hs-source-dirs:    test
   build-depends:     aeson
                    , base       >= 4.12     && < 5
