+++ test/Common/Include.hs
@@ -1,52 +1,17 @@
 -- This template expects CPP definitions for:
 --     PLATFORM_NAME = Posix | Windows
 
-{-# LANGUAGE MultiParamTypeClasses #-}
-{-# LANGUAGE QuasiQuotes #-}
 {-# LANGUAGE TemplateHaskell #-}
-{-# OPTIONS_GHC -Wno-unused-top-binds #-}
 
 -- | Test functions that are common to Posix and Windows
 
 module Common.PLATFORM_NAME (extensionOperations) where
 
 import Control.Monad
-import qualified Language.Haskell.TH.Syntax as TH
 import Path.Internal.PLATFORM_NAME
 import Path.PLATFORM_NAME
 import System.FilePath.PLATFORM_NAME (pathSeparator)
 import Test.Hspec
-
--- | This is a helper type class that checks that splices produce a 'Path' with
---   all type variables instantiated to a type.
---   This ensures that bugs like https://github.com/commercialhaskell/path/issues/159
---   cannot happen.
-class CheckInstantiated a b where
-    checkInstantiated :: Path a b -> FilePath
-    checkInstantiated = toFilePath
-
-instance CheckInstantiated Abs Dir
-instance CheckInstantiated Abs File
-instance CheckInstantiated Rel Dir
-instance CheckInstantiated Rel File
-
-qqRelDir :: FilePath
-qqRelDir = checkInstantiated [reldir|foo/|]
-
-qqRelFile :: FilePath
-qqRelFile = checkInstantiated [relfile|foo|]
-
-thRelDir :: FilePath
-thRelDir = checkInstantiated $(mkRelDir "foo/")
-
-thRelFile :: FilePath
-thRelFile = checkInstantiated $(mkRelFile "foo")
-
-liftRelDir :: FilePath
-liftRelDir = checkInstantiated $(TH.lift (Path "foo/" :: Path Rel Dir))
-
-liftRelFile :: FilePath
-liftRelFile = checkInstantiated $(TH.lift (Path "foo" :: Path Rel File))
 
 validExtensionsSpec :: String -> Path b File -> Path b File -> Spec
 validExtensionsSpec ext file fext = do
